requirements.txt
Django==4.1.4
python-dotenv==1.0.0
requests==2.28.1
